refactor: getQuestionUnits を QuestionGroupPanel に移動
