+++ src/components/information/QuestionGroupPanel.vue
@@ -93,6 +93,37 @@ const refreshAnswersMap = async () => {
   isReady.value = true
 }
 
+// 質問配列を 1 列 / 2 列ユニットに整形する純粋関数
+const getQuestionUnits = (questions: Question[]): Array<{ size: 1 | 2; questions: Question[] }> => {
+  const units: { size: 1 | 2; questions: Question[] }[] = []
+  for (const question of questions) {
+    if (question.type === 'free_text' || question.type === 'multiple') {
+      units.push({ size: 2, questions: [question] })
+    } else {
+      if (units.length === 0 || units[units.length - 1].size === 2) {
+        units.push({ size: 1, questions: [question] })
+      } else {
+        units[units.length - 1].questions.push(question)
+      }
+    }
+  }
+
+  // size 1 のユニットが奇数個のときは最後の 1 つを 2 列にして余白を埋める
+  for (let i = 0; i < units.length; i++) {
+    if (units[i].size === 1 && units[i].questions.length % 2 === 1) {
+      const q = units[i].questions.pop()!
+      units.splice(i + 1, 0, { size: 2, questions: [q] })
+      if (units[i].questions.length === 0) {
+        units.splice(i, 1)
+      } else {
+        i++
+      }
+    }
+  }
+
+  return units
+}
+
 // すべての質問に適切な回答が与えられているか
 const allChecked = computed(() => {
   let result = true
@@ -150,7 +181,7 @@ const sendAnswers = async () => {
   }
 
   if (isAnswered.value) {
-    // 変更された回答のPUTリクエストを並列で実行
+    // 変更された回答の PUT リクエストを並列で実行
     const updatePromises = props.questionGroup.questions
       .filter((question: Question) => isAnswerChanged(question.id))
       .map(async (question) => {
@@ -190,6 +221,7 @@ onMounted(refreshAnswersMap)
     :answers-map="answersMap"
     :all-checked="allChecked"
     :is-answered="isAnswered"
+    :get-question-units="getQuestionUnits"
     @update:answer="({ questionId, value }) => (answersMap[questionId].value = value)"
     @save="sendAnswers"
     @close="quitEditMode"
@@ -200,6 +232,7 @@ onMounted(refreshAnswersMap)
     :answers-map="answersMap"
     :is-ready="isReady"
     :is-editable="isEditable"
+    :get-question-units="getQuestionUnits"
     @edit="inEditMode = true"
   />
 </template>
